Make cull dockerfile agnostic to URLs
In general, dockerfiles should be generic for use everywhere,
and not specific to a particular installation. We can achieve
this by not putting IPs etc in them - and instead passing them
in via environment variables like this. Kubernetes creates
these environment variables for us for each service for exactly
this purpose.

We call the command via a shell so we can do environment variable
expansion.