Provide a JupyterHub token for cull service via a ConfigMap
We create a secret value that'll act as our JupyterHub API token,
and then we provide it to both the hub and the culler via environment
variables bound from the ConfigMap. Thiis is similar to how we share
CONFIGPROXY_AUTH_TOKEN between hub and proxy in the separate proxy
setups.